tutorials/google-adk/07_callbacks/pyproject.toml
[project]
name = "07-callbacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
]
